Refactor string utilities
