+++ src/webview/modules/stateManager.js
@@ -20,7 +20,7 @@ import {
   safeSetElementValue,
   safeSetElementChecked,
 } from './utils.js';
-import { capitalize } from './utils.js';
+import { capitalize } from './stringUtils.js';
 
 export function setDefaultState() {
   // Hide output name override by default
