+++ src/webview/modules/stringUtils.js
@@ -1,0 +1,7 @@
+export function capitalize(str) {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
+
+export function uncapitalize(str) {
+  return str.charAt(0).toLowerCase() + str.slice(1);
+}
