+++ src/webview/modules/utils.js
@@ -1,3 +1,5 @@
+export { capitalize, uncapitalize } from './stringUtils.js';
+
 export function addEventListenerSafely(elementOrId, event, handler) {
   const element =
     typeof elementOrId === 'string'
@@ -51,12 +53,4 @@ export function safeGetElementChecked(id) {
     return false;
   }
   return element.checked;
-}
-
-export function capitalize(str) {
-  return str.charAt(0).toUpperCase() + str.slice(1);
-}
-
-export function uncapitalize(str) {
-  return str.charAt(0).toLowerCase() + str.slice(1);
 }
